Move all consumer user crud operations to a single place
